Create 16 by 16 grid of squares
The etch-a-sketch is represented by a 16 by 16 grid of squares.
These squares have been created dynamically in the JavaScript
file rather than the html file. The grid is created with nested
flexboxes. For now, each grid square is given a random color as
a background in order to visualize the structure of the grid.