Add hover effect to allow drawing on grid
When the mouse passes over a square on the grid, the color of the
square needs to change so that the user can "draw" on it. This has
been implemented by adding an event listener to each square that
fires on the "mousenter" event. For now, the color of the square is
changed to a random rgb color value.