Add functionality to change grid size
Before, the grid size could not be changed. Now, there is a
button at the top of the webpage that allows the user to change
the grid size. When it is clicked, a popup asks the user to enter
a number of squares per side for the new grid. If a valid number
is entered, the current squares in the grid are removed and the
grid is rebuilt. The limit for the size of the grid side is 100
squares to prevent the webpage from becoming too slow.