Finish styling of webpage
Before, the webpage had a few basic styiling details, but was
mainly just functional. Now, there are a few basic styling
elements added to the site. The title 'Etch-A-Sketch' has a
custom font that matches the theme of the game, the options
column has been given more spacing, and the drawing grid has
been given a border and a shadow effect.